+++ mmaction/datasets/registry.py
@@ -2,3 +2,4 @@ from mmcv.utils import Registry
 
 DATASETS = Registry('dataset')
 PIPELINES = Registry('pipeline')
+BLENDINGS = Registry('blending')